+++ src/LiveComponent/src/EventListener/LiveComponentSubscriber.php
@@ -13,7 +13,6 @@ namespace Symfony\UX\LiveComponent\EventListener;
 
 use Psr\Container\ContainerInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Event\ControllerEvent;
@@ -29,6 +28,7 @@ use Symfony\Component\Security\Csrf\CsrfToken;
 use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 use Symfony\Contracts\Service\ServiceSubscriberInterface;
 use Symfony\UX\LiveComponent\Attribute\AsLiveComponent;
+use Symfony\UX\LiveComponent\Attribute\LiveArg;
 use Symfony\UX\LiveComponent\LiveComponentHydrator;
 use Symfony\UX\TwigComponent\ComponentFactory;
 use Symfony\UX\TwigComponent\ComponentRenderer;
@@ -41,8 +41,7 @@ use Symfony\UX\TwigComponent\ComponentRenderer;
  */
 class LiveComponentSubscriber implements EventSubscriberInterface, ServiceSubscriberInterface
 {
-    private const JSON_FORMAT = 'live-component-json';
-    private const JSON_CONTENT_TYPE = 'application/vnd.live-component+json';
+    private const HTML_CONTENT_TYPE = 'application/vnd.live-component+html';
 
     private ContainerInterface $container;
 
@@ -68,8 +67,6 @@ class LiveComponentSubscriber implements EventSubscriberInterface, ServiceSubscr
         if (!$this->isLiveComponentRequest($request)) {
             return;
         }
-
-        $request->setFormat(self::JSON_FORMAT, self::JSON_CONTENT_TYPE);
 
         // the default "action" is get, which does nothing
         $action = $request->get('action', 'get');
@@ -141,11 +138,21 @@ class LiveComponentSubscriber implements EventSubscriberInterface, ServiceSubscr
 
         $this->container->get(LiveComponentHydrator::class)->hydrate($component, $data);
 
+        $request->attributes->set('_component', $component);
+
+        if (!\is_string($queryString = $request->query->get('args'))) {
+            return;
+        }
+
         // extra variables to be made available to the controller
         // (for "actions" only)
-        parse_str($request->query->get('values'), $values);
-        $request->attributes->add($values);
-        $request->attributes->set('_component', $component);
+        parse_str($queryString, $args);
+
+        foreach (LiveArg::liveArgs($component, $action) as $parameter => $arg) {
+            if (isset($args[$arg])) {
+                $request->attributes->set($parameter, $args[$arg]);
+            }
+        }
     }
 
     public function onKernelView(ViewEvent $event): void
@@ -192,7 +199,7 @@ class LiveComponentSubscriber implements EventSubscriberInterface, ServiceSubscr
             return;
         }
 
-        if (!$this->isLiveComponentJsonRequest($request)) {
+        if (!\in_array(self::HTML_CONTENT_TYPE, $request->getAcceptableContentTypes(), true)) {
             return;
         }
 
@@ -200,8 +207,9 @@ class LiveComponentSubscriber implements EventSubscriberInterface, ServiceSubscr
             return;
         }
 
-        $event->setResponse(new JsonResponse([
-            'redirect_url' => $response->headers->get('Location'),
+        $event->setResponse(new Response(null, 204, [
+            'Location' => $response->headers->get('Location'),
+            'Content-Type' => self::HTML_CONTENT_TYPE,
         ]));
     }
 
@@ -227,27 +235,11 @@ class LiveComponentSubscriber implements EventSubscriberInterface, ServiceSubscr
             $request->attributes->get('_component_template')
         );
 
-        if ($this->isLiveComponentJsonRequest($request)) {
-            return new JsonResponse(
-                [
-                    'html' => $html,
-                    'data' => $this->container->get(LiveComponentHydrator::class)->dehydrate($component),
-                ],
-                200,
-                ['Content-Type' => self::JSON_CONTENT_TYPE]
-            );
-        }
-
         return new Response($html);
     }
 
     private function isLiveComponentRequest(Request $request): bool
     {
         return 'live_component' === $request->attributes->get('_route');
-    }
-
-    private function isLiveComponentJsonRequest(Request $request): bool
-    {
-        return \in_array($request->getPreferredFormat(), [self::JSON_FORMAT, 'json'], true);
     }
 }